[REFACTOR] 코드 리팩토링 - 코드 수정 - getResizePicture() ImageUtil로 이동. feature/#61
